+++ app/Forms/Types/ItemType.php
@@ -20,6 +20,11 @@ class ItemType extends AbstractType
             'allow_add' => true,
             'allow_delete' => true,
             'by_reference' => false,
+            'prototype' => true,
+            'attr' => [
+                // @todo move to controller/view
+                'class' => 'js-form-collection',
+            ]
         ]);
     }