+++ resources/views/form/default/checkbox_widget.blade.php
@@ -1,0 +1,3 @@
+<input type="checkbox" @include('form.default.widget_attributes')
+@if (isset($value)) value="{{ $value }}"@endif
+@if ($checked) checked="checked"@endif />