+++ resources/views/form/default/form_widget_simple.blade.php
@@ -1,0 +1,2 @@
+@php $type = isset($type) ? $type : 'text' @endphp
+<input type="{{ $type }}" @include('form.default.widget_attributes') @if (!empty($value))value="{{ $value }}" @endif/>