+++ resources/views/form/default/submit_widget.blade.php
@@ -1,0 +1,2 @@
+@php $type = isset($type) ? $type : 'submit' @endphp
+@include('form.default.button_widget', ['type' => $type])